+++ apps/web/src/components/AdPanel/config.tsx
@@ -1,17 +1,15 @@
 import { useMatchBreakpoints } from '@pancakeswap/uikit'
-import { atom, useAtomValue } from 'jotai'
-import { useMemo } from 'react'
 import { AdsIds, useAdsConfigs } from 'components/AdPanel/hooks/useAdsConfig'
+import { useMemo } from 'react'
 import { AdCakeStaking } from './Ads/AdCakeStaking'
+import { AdCommon } from './Ads/AdCommon'
 import { AdIfo } from './Ads/AdIfo'
 import { AdPCSX } from './Ads/AdPCSX'
-import { AdPicks } from './Ads/AdPicks'
 import { AdSpringboard } from './Ads/AdSpringboard'
 import { ExpandableAd } from './Expandable/ExpandableAd'
 import { shouldRenderOnPages } from './renderConditions'
-import { AdSlide, PicksConfig } from './types'
+import { AdSlide } from './types'
 import { useShouldRenderAdIfo } from './useShouldRenderAdIfo'
-import { AdCommon } from './Ads/AdCommon'
 
 enum Priority {
   FIRST_AD = 6,
@@ -20,36 +18,6 @@ enum Priority {
   MEDIUM = 3,
   LOW = 2,
   VERY_LOW = 1,
-}
-
-const picksConfigAtom = atom(async () => {
-  const time = Math.floor((Date.now() / 1000) * 60 * 5) // Cache 5min
-
-  const urlPreview = `https://proofs.pancakeswap.com/picks/today-preview.json?t=${time}`
-  const isPreview = window.location.origin !== 'https://pancakeswap.finance'
-  const url = isPreview ? urlPreview : `https://proofs.pancakeswap.com/picks/today.json?t=${time}`
-  try {
-    const response = await fetch(url)
-    const json = await response.json()
-    return json as PicksConfig
-  } catch (ex) {
-    return null
-  }
-})
-export const usePicksConfig = () => {
-  const picksConfig = useAtomValue(picksConfigAtom)
-
-  if (!picksConfig) {
-    return []
-  }
-
-  const adList: AdSlide[] = picksConfig.configs.map((config, i) => {
-    return {
-      id: `pick-${config.poolId}`,
-      component: <AdPicks config={config} index={i} />,
-    }
-  })
-  return adList
 }
 
 export const useAdConfig = () => {
