+++ apps/web/src/components/AdPanel/hooks/usePicksConfig.tsx
@@ -1,0 +1,33 @@
+import { atom, useAtomValue } from 'jotai'
+import { AdPicks } from '../Ads/AdPicks'
+import { AdSlide, PicksConfig } from '../types'
+
+const picksConfigAtom = atom(async () => {
+  const time = Math.floor((Date.now() / 1000) * 60 * 5) // Cache 5min
+
+  const urlPreview = `https://proofs.pancakeswap.com/picks/today-preview.json?t=${time}`
+  const isPreview = window.location.origin !== 'https://pancakeswap.finance'
+  const url = isPreview ? urlPreview : `https://proofs.pancakeswap.com/picks/today.json?t=${time}`
+  try {
+    const response = await fetch(url)
+    const json = await response.json()
+    return json as PicksConfig
+  } catch (ex) {
+    return null
+  }
+})
+export const usePicksConfig = () => {
+  const picksConfig = useAtomValue(picksConfigAtom)
+
+  if (!picksConfig) {
+    return []
+  }
+
+  const adList: AdSlide[] = picksConfig.configs.map((config, i) => {
+    return {
+      id: `pick-${config.poolId}`,
+      component: <AdPicks config={config} index={i} />,
+    }
+  })
+  return adList
+}
