+++ apps/web/src/views/universalFarms/components/PoolsBanner.tsx
@@ -2,8 +2,6 @@ import { useTheme } from '@pancakeswap/hooks'
 import { useTranslation } from '@pancakeswap/localization'
 import { Box, Button, Column, LinkExternal, PageHeader, Row, Text } from '@pancakeswap/uikit'
 import { VerticalDivider } from '@pancakeswap/widgets-internal'
-import { PickAdSlides } from 'components/AdPanel/PickAdSlides'
-import { Suspense } from 'react'
 import { FarmFlexWrapper, FarmH1, FarmH2 } from 'views/Farms/styled'
 
 export const PoolsBanner = ({ additionLink }: { additionLink?: React.ReactNode }) => {
@@ -39,11 +37,6 @@ export const PoolsBanner = ({ additionLink }: { additionLink?: React.ReactNode }
                 </>
               )}
             </Row>
-          </Box>
-          <Box>
-            <Suspense>
-              <PickAdSlides isDismissible={false} />
-            </Suspense>
           </Box>
         </FarmFlexWrapper>
       </Column>
